Moved loggin stuff from BlossomsPoGoManager to AccountController
First cleanup of the main application class
