+++ src/me/corriekay/pokegoutil/windows/MenuBar.java
@@ -7,6 +7,7 @@ import com.pokegoapi.api.player.PlayerProfile;
 import com.pokegoapi.api.player.PlayerProfile.Currency;
 
 import me.corriekay.pokegoutil.BlossomsPoGoManager;
+import me.corriekay.pokegoutil.controllers.AccountController;
 
 @SuppressWarnings("serial")
 public class MenuBar extends JMenuBar {
@@ -54,7 +55,7 @@ public class MenuBar extends JMenuBar {
 	}
 	
 	private void logout() throws Exception {
-		BlossomsPoGoManager.logOff();
+		AccountController.logOff();
 	}
 
 	private void displayTrainerStats() throws Exception {
