+++ src/me/corriekay/pokegoutil/windows/PokemonTab.java
@@ -8,7 +8,6 @@ import java.time.ZoneId;
 import java.util.*;
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
 
 import javax.swing.*;
 import javax.swing.RowSorter.SortKey;
